requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
Werkzeug==1.0.1
